fixed issue #304

+++ packages/Webkul/User/src/Http/Controllers/UserController.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Facades\Event;
 use Webkul\User\Repositories\AdminRepository as Admin;
 use Webkul\User\Repositories\RoleRepository as Role;
 use Webkul\User\Http\Requests\UserForm;
+use Hash;
 
 /**
  * Admin user controller
@@ -94,7 +95,7 @@ class UserController extends Controller
             $data['password'] = bcrypt($data['password']);
 
         Event::fire('user.admin.create.before');
-        
+
         $admin = $this->admin->create($data);
 
         Event::fire('user.admin.delete.after', $admin);
@@ -165,6 +166,10 @@ class UserController extends Controller
         } else {
             Event::fire('user.admin.delete.before', $id);
 
+            if (auth()->guard('admin')->user()->id == $id) {
+                return view('admin::customers.confirm-password');
+            }
+
             $this->admin->delete($id);
 
             Event::fire('user.admin.delete.after', $id);
@@ -173,5 +178,37 @@ class UserController extends Controller
         }
 
         return redirect()->back();
+    }
+
+    /**
+     * destroy current after confirming
+     *
+     * @return mixed
+     */
+    public function destroySelf()
+    {
+        $password = request()->input('password');
+
+        if(Hash::check($password, auth()->guard('admin')->user()->password)) {
+            if($this->admin->count() == 1) {
+                session()->flash('error', trans('admin::app.users.users.delete-last'));
+            } else {
+                $id = auth()->guard('admin')->user()->id;
+
+                Event::fire('user.admin.delete.before', $id);
+
+                $this->admin->delete($id);
+
+                Event::fire('user.admin.delete.after', $id);
+
+                session()->flash('success', trans('admin::app.users.users.delete-success'));
+
+                return redirect()->route('admin.session.create');
+            }
+        } else {
+            session()->flash('warning', trans('admin::app.users.users.incorrect-password'));
+
+            return redirect()->route($this->_config['redirect']);
+        }
     }
 }
